refactor the page scss so it isn't included in the sitemap
